requirements.txt
# Core ML and NLP dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
scikit-learn>=1.0.0

# LangGraph and LangChain dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0

# API and HTTP dependencies
requests>=2.28.0
aiohttp>=3.8.0
python-dotenv>=0.19.0

# Data processing and utilities
pandas>=1.5.0
json5>=0.9.0
datasets>=2.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Graph and database (optional)
networkx>=2.8.0
neo4j>=5.0.0
redis>=4.3.0

# Configuration and validation
pydantic>=1.10.0
jsonschema>=4.4.0

# Logging and monitoring
wandb>=0.15.0
tqdm>=4.64.0
rich>=12.0.0
loguru>=0.6.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0

# Type hints
typing-extensions>=4.0.0